tests: Refactor code to address link-time errors.
Dropping the "-O2" compiler flag when code coverage is enabled caused
the inline functions `test-plugin.h' to be unresolved at link-time.
Refactor those functions to a new `libmptcpd_test' convenience library
to make them out-of-line, allowing the code to link when both
optimization is enabled and disabled.
