+++ tests/test-plugin.h
@@ -10,15 +10,14 @@
 #ifndef MPTCPD_TEST_PLUGIN_H
 #define MPTCPD_TEST_PLUGIN_H
 
-#include <string.h>
+#include <stdbool.h>
 
 #include <mptcpd/types.h>
-#include <mptcpd/plugin.h>
+
 
 #ifdef __cplusplus
 extern "C" {
 #endif
-
 
 // ------------------------------------------------------------------
 
@@ -51,58 +50,26 @@ struct plugin_call_count
         //@}
 };
 
-inline void plugin_call_count_reset(struct plugin_call_count *p)
-{
-        p->new_connection         = 0;
-        p->connection_established = 0;
-        p->connection_closed      = 0;
-        p->new_address            = 0;
-        p->address_removed        = 0;
-        p->new_subflow            = 0;
-        p->subflow_closed         = 0;
-        p->subflow_priority       = 0;
-}
+/**
+ * @brief Reset plugin operation call counts.
+ */
+void call_count_reset(struct plugin_call_count *p);
 
-inline bool plugin_call_counts_are_pos(struct plugin_call_count const *p)
-{
-        return     p->new_connection >= 0
-                && p->connection_established >= 0
-                && p->connection_closed >= 0
-                && p->new_address >= 0
-                && p->address_removed >= 0
-                && p->new_subflow >= 0
-                && p->subflow_closed >= 0
-                && p->subflow_priority >= 0;
-}
+/**
+ * @brief Check if all plugin operation call counts are not negative.
+ */
+bool call_count_all_positive(struct plugin_call_count const *p);
 
-inline bool plugin_call_count_is_sane(struct plugin_call_count const *p)
-{
-        return  // non-negative counts
-                plugin_call_counts_are_pos(p)
+/**
+ * @brief Perform a sanity check on the plugin operation call counts.
+ */
+bool call_count_is_sane(struct plugin_call_count const *p);
 
-                /*
-                  Some callbacks should not be called more than
-                  others.
-                */
-                && p->connection_established <= p->new_connection
-                && p->connection_closed <= p->new_connection
-                && p->subflow_closed    <= p->new_subflow;
-}
-
-inline bool plugin_call_count_is_equal(
-        struct plugin_call_count const *lhs,
-        struct plugin_call_count const *rhs)
-{
-        return lhs->new_connection         == rhs->new_connection
-            && lhs->connection_established == rhs->connection_established
-            && lhs->connection_closed      == rhs->connection_closed
-            && lhs->new_address            == rhs->new_address
-            && lhs->address_removed        == rhs->address_removed
-            && lhs->new_subflow            == rhs->new_subflow
-            && lhs->subflow_closed         == rhs->subflow_closed
-            && lhs->subflow_priority       == rhs->subflow_priority;
-
-}
+/**
+ * @brief Compare plugin operation call counts for equality.
+ */
+bool call_count_is_equal(struct plugin_call_count const *lhs,
+                         struct plugin_call_count const *rhs);
 
 // ------------------------------------------------------------------
 
@@ -114,7 +81,7 @@ inline bool plugin_call_count_is_equal(
  * intended to call were actually called.
  */
 //@{
-struct plugin_call_count const test_count_1 = {
+static struct plugin_call_count const test_count_1 = {
         .new_connection         = 1,
         .connection_established = 1,
         .connection_closed      = 1,
@@ -125,7 +92,7 @@ struct plugin_call_count const test_count_1 = {
         .subflow_priority       = 0
 };
 
-struct plugin_call_count const test_count_2 = {
+static struct plugin_call_count const test_count_2 = {
         .new_connection         = 1,
         .connection_established = 1,
         .connection_closed      = 1,
@@ -136,7 +103,7 @@ struct plugin_call_count const test_count_2 = {
         .subflow_priority       = 1
 };
 
-struct plugin_call_count const test_count_4 = {
+static struct plugin_call_count const test_count_4 = {
         .new_connection         = 1,
         .connection_established = 1,
         .connection_closed      = 1,
@@ -148,7 +115,6 @@ struct plugin_call_count const test_count_4 = {
 };
 //@}
 
-
 // ------------------------------------------------------------------
 
 /**
@@ -159,31 +125,31 @@ struct plugin_call_count const test_count_4 = {
  * types of values they correspond to.
  */
 //@{
-mptcpd_token_t const test_token_1    = 0x12345678;
-mptcpd_aid_t   const test_laddr_id_1 = 0x34;
-mptcpd_aid_t   const test_raddr_id_1 = 0x56;
-bool           const test_backup_1   = true;
+static mptcpd_token_t const test_token_1    = 0x12345678;
+static mptcpd_aid_t   const test_laddr_id_1 = 0x34;
+static mptcpd_aid_t   const test_raddr_id_1 = 0x56;
+static bool           const test_backup_1   = true;
 
-mptcpd_token_t const test_token_2    = 0x23456789;
-mptcpd_aid_t   const test_laddr_id_2 = 0x23;
-mptcpd_aid_t   const test_raddr_id_2 = 0x45;
-bool           const test_backup_2   = false;
+static mptcpd_token_t const test_token_2    = 0x23456789;
+static mptcpd_aid_t   const test_laddr_id_2 = 0x23;
+static mptcpd_aid_t   const test_raddr_id_2 = 0x45;
+static bool           const test_backup_2   = false;
 
-mptcpd_token_t const test_token_4    = 0x34567890;
-mptcpd_aid_t   const test_laddr_id_4 = 0x90;
-mptcpd_aid_t   const test_raddr_id_4 = 0x01;
-bool           const test_backup_4   = true;
+static mptcpd_token_t const test_token_4    = 0x34567890;
+static mptcpd_aid_t   const test_laddr_id_4 = 0x90;
+static mptcpd_aid_t   const test_raddr_id_4 = 0x01;
+static bool           const test_backup_4   = true;
 
 // For verifying that a plugin will not be dispatched.
-mptcpd_token_t const test_bad_token  = 0xFFFFFFFF;
+static mptcpd_token_t const test_bad_token  = 0xFFFFFFFF;
 
-struct mptcpd_addr const test_laddr_1 = {
+static struct mptcpd_addr const test_laddr_1 = {
         .address = { .family = AF_INET,
                      .addr   = { .addr4 = { .s_addr = 0x34567890 } } },
         .port    = 0x1234
 };
 
-struct mptcpd_addr const test_laddr_2 = {
+static struct mptcpd_addr const test_laddr_2 = {
 #ifdef __cplusplus
         /*
           s6_addr is a macro that expands to code that isn't
@@ -220,7 +186,7 @@ struct mptcpd_addr const test_laddr_2 = {
         .port    = 0x5678
 };
 
-struct mptcpd_addr const test_raddr_1 = {
+static struct mptcpd_addr const test_raddr_1 = {
 #ifdef __cplusplus
         /*
           s6_addr is a macro that expands to code that isn't
@@ -257,19 +223,19 @@ struct mptcpd_addr const test_raddr_1 = {
         .port    = 0x3456
 };
 
-struct mptcpd_addr const test_raddr_2 = {
+static struct mptcpd_addr const test_raddr_2 = {
         .address = { .family = AF_INET,
                      .addr   = { .addr4 = { .s_addr = 0x98765432 } } },
         .port    = 0x7890
 };
 
-struct mptcpd_addr const test_laddr_4 = {
+static struct mptcpd_addr const test_laddr_4 = {
         .address = { .family = AF_INET,
                      .addr   = { .addr4 = { .s_addr = 0x45678901 } } },
         .port    = 0x2345
 };
 
-struct mptcpd_addr const test_raddr_4 = {
+static struct mptcpd_addr const test_raddr_4 = {
         .address = { .family = AF_INET,
                      .addr   = { .addr4 = { .s_addr = 0x56789012 } } },
         .port    = 0x3456
@@ -279,89 +245,30 @@ struct mptcpd_addr const test_raddr_4 = {
 
 /**
  * @brief Compare equality of two @c mptcpd_addr objects.
+ *
+ * @return @c true if both @c mptcpd_addr objects are equal. @c false
+ *         otherwise.
  */
-extern bool mptcpd_addr_is_equal(struct mptcpd_addr const *lhs,
-                                 struct mptcpd_addr const *rhs)
-{
-        if (lhs->address.family == rhs->address.family
-            && lhs->port == rhs->port) {
-                struct mptcpd_in_addr const *const left =
-                        &lhs->address;
-                struct mptcpd_in_addr const *const right =
-                        &rhs->address;
+bool mptcpd_addr_is_equal(struct mptcpd_addr const *lhs,
+                          struct mptcpd_addr const *rhs);
 
-                if (lhs->address.family == AF_INET)
-                        return left->addr.addr4.s_addr
-                                == right->addr.addr4.s_addr;
-                else
-                        // memcmp() is fine for this case.
-                        return memcmp(
-                                &left->addr.addr6.s6_addr,
-                                &right->addr.addr6.s6_addr,
-                                sizeof(left->addr.addr6.s6_addr)) == 0;
-        }
-
-        return false;
-}
-
-extern void call_plugin_ops(struct plugin_call_count const *count,
-                            char const *name,
-                            mptcpd_token_t token,
-                            mptcpd_aid_t raddr_id,
-                            struct mptcpd_addr const *laddr,
-                            struct mptcpd_addr const *raddr,
-                            bool backup)
-{
-        assert(count != NULL);
-
-        for (int i = 0; i < count->new_connection; ++i)
-                mptcpd_plugin_new_connection(name,
-                                             token,
-                                             laddr,
-                                             raddr,
-                                             NULL);
-
-        for (int i = 0; i < count->connection_established; ++i)
-                mptcpd_plugin_connection_established(token,
-                                                     laddr,
-                                                     raddr,
-                                                     NULL);
-
-        for (int i = 0; i < count->new_address; ++i)
-                mptcpd_plugin_new_address(token,
-                                          raddr_id,
-                                          raddr,
-                                          NULL);
-
-        for (int i = 0; i < count->address_removed; ++i)
-                mptcpd_plugin_address_removed(token,
-                                              raddr_id,
-                                              NULL);
-
-        for (int i = 0; i < count->new_subflow; ++i)
-                mptcpd_plugin_new_subflow(token,
-                                          laddr,
-                                          raddr,
-                                          backup,
-                                          NULL);
-
-        for (int i = 0; i < count->subflow_closed; ++i)
-                mptcpd_plugin_subflow_closed(token,
-                                             laddr,
-                                             raddr,
-                                             backup,
-                                             NULL);
-
-        for (int i = 0; i < count->subflow_priority; ++i)
-                mptcpd_plugin_subflow_priority(token,
-                                               laddr,
-                                               raddr,
-                                               backup,
-                                               NULL);
-
-        for (int i = 0; i < count->connection_closed; ++i)
-                mptcpd_plugin_connection_closed(token, NULL);
-}
+/**
+ * @brief Call plugin operations
+ *
+ * @param[in] count    Number of times to call each plugin operation.
+ * @param[in] name     Plugin name.
+ * @param[in] raddr_id Remote address ID.
+ * @param[in] laddr    Local address.
+ * @param[in] raddr    Remote address.
+ * @param[in] backup   MPTCP backup priority.
+ */
+void call_plugin_ops(struct plugin_call_count const *count,
+                     char const *name,
+                     mptcpd_token_t token,
+                     mptcpd_aid_t raddr_id,
+                     struct mptcpd_addr const *laddr,
+                     struct mptcpd_addr const *raddr,
+                     bool backup);
 
 
 #ifdef __cplusplus
